requirements.txt
azure-functions==1.9.0